WL-2948 http:// loads in new window
This is a squashed commit composed of

Partial fix, the new javascript method works but is loaded when the top level window loads as opposed to when the iframe loads.

Works for iframes of arbitrary depth and IE 8 too. The fix is included on every page, likely overkill.

Changed call to a method to match the rest of headscripts, removed effect on target '_top'

Including check that the link actually has an href

(https://github.com/ox-it/wl-reference/pull/1#discussion_r6152912)