WL-2961 Links in Oxitems feeds open in a new window
This is a combination of 4 commits

Refactored some of the functions for handling the mixed content stuff. This involves a function rename and so needs an update to kernel.

Adds the check that we are in an iframe whenever links are fixed.

Fixed reference to old method.

Oops, previous commit excluded the youtube part of mixed content.