WL-2998 Update to use new podcast feed.
Some browsers require the namespace prefix and some dont, but if our selector just gives both then we don't have to detect which works on each browser.
We now work out the type of the feed based on the category tag.

We also stop converting the XML to a jQuery object multiple times to improve performance.