WL-3089 Re-enable the right mouse context menu in CK
This was disabled becuase in most cases the context menu only provides a
paste button which is much less useful than the browser's native right
mouse menu, especially for things like spellings.

There are some situations where the context menu is actually useful and
there are no alternative areas with the same functionality. For example
when editting a table the context menu provides column and row
manipulations.