WL-1631 Round times consistently.
The date/time picker now allows only 5 minute intervals, but when fields in the evaluations tool are setup they contain date/times that aren’t already rounded. There was some Sakai code to perform initial rounding but this used a different algorithm and would round the minute value across days so then trying to create a time at the end of a day it would jump into the next day. 

The plugin does sensible rounding so the change just gets the rounded value from the picker after we’ve set it up and resets all the hidden field to that value.

It also fixes a bug where if the initial set time wasn’t on a rounded boundary the displayed time didn’t match the value that was submitted in the form.